batch: Move line mapping into matching package
The batch implementation has recursive import checks while its line-mapping
type remains among unrelated flat modules.

The project cannot present matching as a bounded responsibility while its
foundational mapping value remains at the batch package root.

This commit continues the package-layout series by moving the mapping
implementation under line_matching and retaining its selected import path
during the test transition.

Subsequent commits will move the workspace, matching algorithm, comparison
helpers, range views, sequence helpers, and lineage.
