+++ src/git_stage_batch/batch/line_matching/__init__.py
@@ -1,0 +1,1 @@
+"""Line comparison, mapping, matching, and lineage helpers."""
